Move query docs to a separate doc
